# docs/swagger.yaml
basePath: /
definitions:
...
  internal_schedules.CreateScheduleRequest:
    properties:
      arrival_time:
        type: string
      bus_id:
        type: string
      departure_time:
        type: string
    type: object
  internal_schedules.Schedule:
    properties:
      arrival_time:
        type: string
      bus_id:
        type: string
      created_at:
        type: string
      departure_time:
        type: string
      id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: SmartTransit Backend APIs
  title: SmartTransit Bus API
  version: "1.0"
paths:
...
  /api/schedules/:
    get:
      description: Retrieve all schedules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_schedules.Schedule'
            type: array
      summary: Get all schedules
      tags:
      - Schedules
    post:
      consumes:
      - application/json
      description: Create new schedule
      parameters:
      - description: Create Schedule
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/internal_schedules.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create schedule
      tags:
      - Schedules
  /api/schedules/{id}:
    delete:
      description: Delete schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete schedule
      tags:
      - Schedules
    get:
      description: Retrieve single schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_schedules.Schedule'
      summary: Get schedule by ID
      tags:
      - Schedules
    put:
      consumes:
      - application/json
      description: Update existing schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Schedule
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/internal_schedules.Schedule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update schedule
      tags:
      - Schedules
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"